+++ test/unit/specs/elements/GeoDropdown/GeoDropdownRegularButton.spec.js
@@ -19,13 +19,13 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container__icon').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container__icon').exists()).toBe(true)
   })
 
   it('Should not render icon if not given', function () {
     const wrapper = mount(GeoDropdownRegularButton, { })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container__icon').exists()).toBe(false)
+    expect(wrapper.find('.geo-dropdown-regular-button-container__icon').exists()).toBe(false)
   })
 
   it('Should render default slot content when given', async function () {
@@ -36,20 +36,20 @@ describe('GeoDropdownRegularButton', () => {
     })
 
     await Vue.nextTick()
-    expect(wrapper.find('.geo-dropdown__regular-button-container__content .my-demo-content').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container__content .my-demo-content').exists()).toBe(true)
   })
 
   it('Should not render container if default slot if empty', async function () {
     const wrapper = mount(GeoDropdownRegularButton, { })
 
     await Vue.nextTick()
-    expect(wrapper.find('.geo-dropdown__regular-button-container__content').exists()).toBe(false)
+    expect(wrapper.find('.geo-dropdown-regular-button-container__content').exists()).toBe(false)
   })
 
   it('Should emit click event when clicked', function () {
     const wrapper = mount(GeoDropdownRegularButton, { })
 
-    wrapper.find('.geo-dropdown__regular-button-container').trigger('click')
+    wrapper.find('.geo-dropdown-regular-button-container').trigger('click')
     expect(wrapper.emitted().click).toBeTruthy()
   })
 
@@ -60,7 +60,7 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    wrapper.find('.geo-dropdown__regular-button-container').trigger('click')
+    wrapper.find('.geo-dropdown-regular-button-container').trigger('click')
     expect(wrapper.emitted().click).toBeFalsy()
   })
 
@@ -71,8 +71,8 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container--active').exists()).toBe(true)
-    expect(wrapper.find('.geo-dropdown__regular-button-container--disabled').exists()).toBe(false)
+    expect(wrapper.find('.geo-dropdown-regular-button-container--active').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container--disabled').exists()).toBe(false)
   })
 
   it('Should apply proper class when disabled', function () {
@@ -82,8 +82,8 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container--active').exists()).toBe(false)
-    expect(wrapper.find('.geo-dropdown__regular-button-container--disabled').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container--active').exists()).toBe(false)
+    expect(wrapper.find('.geo-dropdown-regular-button-container--disabled').exists()).toBe(true)
   })
 
   it('Should apply proper class when active and disabled', function () {
@@ -94,8 +94,8 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container--active').exists()).toBe(true)
-    expect(wrapper.find('.geo-dropdown__regular-button-container--disabled').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container--active').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container--disabled').exists()).toBe(true)
   })
 
   it('Should apply proper class when icon position by default', function () {
@@ -108,7 +108,7 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container__icon--left').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container__icon--left').exists()).toBe(true)
   })
 
   it('Should apply proper class when icon position right', function () {
@@ -122,6 +122,6 @@ describe('GeoDropdownRegularButton', () => {
       }
     })
 
-    expect(wrapper.find('.geo-dropdown__regular-button-container__icon--right').exists()).toBe(true)
+    expect(wrapper.find('.geo-dropdown-regular-button-container__icon--right').exists()).toBe(true)
   })
 })